// packages/contracts/src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        CustomerHistoryPage: {
            items: components["schemas"]["CustomerHistoryItem"][];
            /** Format: int32 */
            page: number | string;
            hasMore: boolean;
            /** Format: int64 */
            loyaltyPoints: number | string;
            /** @default true */
            canForgetAccount: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}